Fixed bug when passing a matrix to transformPointCloud.
The existing code uses the test "if not Y" to check if a matrix of points has been passed to the funciton. However, this syntax throws an error when Y is neither "None" or a Boolean.

Changed test to "if Y is None", which tests specifically for the default case of transformPointCloud(Y=None)