Deformable low_rank fixing typos, add testing and examples
Fixed final typos and use of old functions in new implementation of low_rank. Added new test that implements the low_rank version of deformable to ensure it is running properly. Also added fish_deformable_3D_lowrank.py example to allow visualization of the low_rank version and to ensure it is working properly.