gaussian_kernel - Allow X & Y instead of using Y twice
Making default paramter inputted to be X and giving the option to specifiy a second dataset Y. If Y is not defined () then make Y=X. This allows for the creation of new gaussian kernels necessary for transforming point clouds that are a different size than the original dataset used for the registration. This is extremely useful as it will allow for registration of subset of the data and then transformation of the original (potentially very large) point cloud. This is an additional way in which things may be sped up (as well as low-rank). The other benefit is that we can transform points/objects that may not have been beneficial/useful in the registration process