transform_point_cloud - Recalculate G if function called externally
Updated the transform_point_cloud so that  it will automatically re-calculate gaussian kernel G. This is done to enable a new gaussian to be created that is specific to the data that is being transformed. This allows for transformation of point clouds that are not the same shape as the original point cloud that was registered. See previous commit ('gaussian_kernel - Allow X & Y instead of using Y twice') for additional description