Simplified gaussian_kernel syntax & added new deformable test
Removed reshape() and tile() commands which can be reduced to one line of directly subtracting Y from X. None is placed in the indices of X/Y on the new axis which that variable should add a dimension of length 1. The new syntax () results in diff being shaped i,j,k where /n This syntax ensures that both X and Y have the same number of dimensions which is equal to  and if they do not an error with be thrown. However, if X and Y don't have the same number of points it will make a rectangular gaussian_kernal () that can be used for transforming point clouds that are different sizes than the original cloud used for registration